Fixed settings hide
