+++ Assets/com.alelievr.NodeGraphProcessor/Editor/Views/ExposedParameterWorkaround.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+using System;
+using System.Collections.Generic;
+
+namespace GraphProcessor
+{
+    [Serializable]
+    public class ExposedParameterWorkaround : ScriptableObject
+    {
+        [SerializeReference]
+        public List<ExposedParameter>   parameters = new List<ExposedParameter>();
+        public BaseGraph                graph;
+    }
+}
