+++ Assets/com.alelievr.NodeGraphProcessor/Editor/Views/ExposedParameterWorkaround.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 886a707ace40af74481202209244b94d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
